refactor: 생성자 파라미터명 XXXcreateData -> XXXcreateCommand 로 변경
