+++ apps/commerce-api/src/main/java/com/loopers/domain/user/User.java
@@ -1,7 +1,7 @@
 package com.loopers.domain.user;
 
 import com.loopers.domain.BaseEntity;
-import com.loopers.domain.user.dto.data.UserCreateCommand;
+import com.loopers.domain.user.dto.command.UserCreateCommand;
 import com.loopers.domain.user.type.GenderType;
 import com.loopers.support.validation.DateConverter;
 import jakarta.persistence.*;
