+++ apps/commerce-api/src/test/java/com/loopers/domain/user/UserTest.java
@@ -1,6 +1,6 @@
 package com.loopers.domain.user;
 
-import com.loopers.domain.user.dto.data.UserCreateCommand;
+import com.loopers.domain.user.dto.command.UserCreateCommand;
 import com.loopers.domain.user.type.GenderType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
